+++ src/data/data.js
@@ -1,0 +1,156 @@
+const data = [
+    {
+            "id": 1,
+            "name_sort": "Fuda :Tokyo",
+            "name": "Fuda :Tokyo Revengers Manji Uniform DTF Print Men & Women Cotton Anime Printed Hoodies",
+            "image": "product4.jpg",
+            "price": 785,
+            "rating": 4.8,
+            "reviews": 1500
+        },
+        {
+            "id": 2,
+            "name_sort": "DUDEME Tokyo",
+            "name": "DUDEME Tokyo Revengers Manji Uniform DTF Print Unisex Anime Hoodie for Men & Women 100% Cotton Anime Printed Regular Fit Hoodies",
+            "image": "product2.jpg",
+            "price": 785,
+            "rating": 4.8,
+            "reviews": 1500
+        },
+        {
+            "id": 3,
+            "name_sort": "DUDEME Tokyo Walhalla",
+            "name": "DUDEME Tokyo Revengers Walhalla Unisex Anime Hoodie for Men & Women 100% Cotton Naruto, Jujutsu Kaisen, Goku, Gojo, Bleach Anime Printed",
+            "image": "product3.jpg",
+            "price": 785,
+            "rating": 4.8,
+            "reviews": 1500
+        },
+        {
+            "id": 4,
+            "name":"Attack on Titan Season 2- Wings of Freedom Wristband",
+            "name_sort":"Attack on Titan ",
+            "image": "product1.jpg",
+            "price": 112,
+            "rating": 4.8,
+            "reviews": 1500
+        },
+        {
+            "id": 5,
+            "name":"TOPLOT Unisex Woolen Blend Hooded Neck Hoodie (5147)",
+            "name_sort":"TOPLOT Unisex  Hoodie",
+            "image": "product5.jpg",
+            "price": 559,
+            "rating": 4.8,
+            "reviews": 1500
+        },
+        {
+            "id": 6,
+            "name":"Lymio Hoodie|| Sweatshirt for Unisex || Unisex Hoodie and Sweatshirt (H-46-47)",
+            "name_sort":"Lymio Hoodie",
+            "image": "product6.jpg",
+            "price": 559,
+            "rating": 4.8,
+            "reviews": 1500
+        },
+        {
+            "id": 7,
+            "name":"Alan Jones Clothing Boys Oversize Hoodie with Graffiti Print | Full Sleeve Sweatshirt for Kids & Teens | Age 6-16 Years",
+            "name_sort":"Alan Jones Hoodie",
+            "image": "product7.jpg",
+            "price": 559,
+            "rating": 4.8,
+            "reviews": 1500
+        },
+        {
+            "id": 8,
+            "name":"Lymio Hoodie|| Sweatshirt for Unisex || Unisex Hoodie and Sweatshirt (H-29)",
+            "name_sort":"Lymio Hoodie",
+            "image": "product8.jpg",
+            "price": 559,
+            "rating": 4.8,
+            "reviews": 1500
+        },
+        {
+            "id": 9,
+            "name":"iBELL FM720BT Portable Pocket FM Radio with Bluetooth Speaker, USB/SD MP3 Player, Rechargeable with Type C Port, 2 Band (Black)",
+            "name_sort":"iBELL FM720BT",
+            "image": "product9.jpg",
+            "price": 112,
+            "rating": 4.8,
+            "reviews": 1500
+        },
+        {
+            "id": 10,
+            "name":"FEDUS 12v Charger for Kids Ride on Car Toys, 12 Volt Battery Charger with Charging Indicator & Auto-Cut Off Battery Power Adaptor for Kids Electric Car Children Mercedes, BMW, Audi, Motorcycle, Jeep",
+            "name_sort":"FEDUS 12v Charger",
+            "image": "product10.jpg",
+            "price": 112,
+            "rating": 4.8,
+            "reviews": 1500
+        },
+        {
+            "id": 11,
+            "name":"pTron Keyvo KW4 Wireless Keyboard & Mouse Combo Set, Ultra-Slim Keyboard w/ 78 Keys, Multimedia Keys, BT Mouse w/ 1600 DPI Precision Optical Sensor, Compatible with Tablets/Smartphones/PC (Black)",
+            "name_sort":"pTron Keyvo KW4",
+            "image": "product11.jpg",
+            "price": 112,
+            "rating": 4.8,
+            "reviews": 1500
+        },
+        {
+            "id": 12,
+            "name":"Aristocrat Nova 15L Laptop Backpack for Men & Women with Bottle Pocket | Padded Shoulder Straps, Multi Compartments | Travel & College Bag | Dark Black",
+            "name_sort":"Aristocrat Nova 15L",
+            "image": "product12.jpg",
+            "price": 112,
+            "rating": 4.8,
+            "reviews": 1500
+        },
+        
+    ]
+
+const sliders = [
+    {
+        id: 1001,
+        image: "slide-p-1.jpg",
+    },
+    {
+        id: 1002,
+        image: "slide-p-2.jpg",
+    },
+    {
+        id: 1003,
+        image: "slide-p-3.jpg",
+    },
+    {
+        id: 1004,
+        image: "slide-p-4.jpg",
+    },
+    {
+        id: 1005,
+        image: "slide-p-5.jpg",
+    },
+    {
+        id: 1006,
+        image: "slide-p-6.jpg",
+    },
+    {
+        id: 1007,
+        image: "slide-p-7.jpg",
+    },
+    {
+        id: 1008,
+        image: "slide-p-8.jpg",
+    },
+   {
+        id: 1009,
+        image: "slide-p-9.jpg",
+   },
+   {
+        id: 1010,
+        image: "slide-p-10.jpg",
+   }
+]
+
+export { data, sliders }